extracted common testing utilities to FormattingTestsBase superclass
