+++ src/test/java/org/docx4j/model/fields/FormattingSwitchHelperInvalidNumericTests.java
@@ -16,7 +16,7 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 @RunWith(value = Parameterized.class)
-public class FormattingSwitchHelperInvalidNumericTests {
+public class FormattingSwitchHelperInvalidNumericTests extends FormattingTestsBase{
 
 	static Locale initialLocale;
 
@@ -70,38 +70,6 @@ public class FormattingSwitchHelperInvalidNumericTests {
 		} catch (Exception e) {
 			System.out.println("[" + data.toString() + "] " + e.getMessage());
 			assertTrue(FieldFormattingException.class.isAssignableFrom(e.getClass()));
-		}
-	}
-
-	// ---------------------------------------------------------------------------------------
-
-	private void doit(String fieldname, SwitchTestData triple, String expectedResult)  throws TransformerException, Docx4JException {
-
-		String instr = fieldname + " foo " + triple.format;
-		String result = getFormat(instr, triple.val);
-		assertTrue(result.equals(expectedResult));
-	}
-
-	private String getFormat(String instr, String val) throws TransformerException, Docx4JException {
-
-		FldSimpleModel fsm = new FldSimpleModel();
-		fsm.build(instr);
-		return FormattingSwitchHelper.applyFormattingSwitch(null, fsm, val);
-	}
-
-	private static class SwitchTestData {
-
-		String format;
-		String val;
-
-		public String toString() {
-			return "format " + format + " to data " + val;
-		}
-
-		public SwitchTestData(String format, String val) {
-
-			this.format = format;
-			this.val = val;
 		}
 	}
