+++ src/test/java/org/docx4j/model/fields/FormattingSwitchHelperNoSwitchTests.java
@@ -13,7 +13,7 @@ import java.util.Arrays;
 import java.util.Collection;
 
 @RunWith(value = Parameterized.class)
-public class FormattingSwitchHelperNoSwitchTests {
+public class FormattingSwitchHelperNoSwitchTests extends FormattingTestsBase{
 
 	// In all cases, output = input
 
@@ -93,39 +93,4 @@ public class FormattingSwitchHelperNoSwitchTests {
 		doit("MERGEFIELD", data, value);
 		doit("DOCPROPERTY", data, value);
 	}
-
-
-	// ---------------------------------------------------------------------------------------
-
-	private void doit(String fieldname, SwitchTestData triple, String expectedResult)  throws TransformerException, Docx4JException {
-
-		String instr = fieldname + " foo " + triple.format;
-		String result = getFormat(instr, triple.val);
-		assertTrue(result.equals(expectedResult));
-	}
-
-	private String getFormat(String instr, String val) throws TransformerException, Docx4JException {
-
-		FldSimpleModel fsm = new FldSimpleModel();
-		fsm.build(instr);
-		return FormattingSwitchHelper.applyFormattingSwitch(null, fsm, val);
-	}
-
-	private static class SwitchTestData {
-
-		String format;
-		String val;
-
-		public String toString() {
-			return "format " + format + " to data " + val;
-		}
-
-		public SwitchTestData(String format, String val) {
-
-			this.format = format;
-			this.val = val;
-		}
-	}
-
-
 }
